requirements.txt
asgiref==3.6.0
boto3==1.26.56
botocore==1.29.56
Django==4.1.5
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-filter==22.1
django-js-asset==2.0.0
django-resized==1.0.2
django-storages==1.13.2
django-tinymce==3.5.0
djangorestframework==3.14.0
jmespath==1.0.1
Markdown==3.4.1
Pillow==9.4.0
psycopg2==2.9.5
python-dateutil==2.8.2
pytz==2022.7.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
